Move back some files to the base folder. We don't need this breaking change.
