+++ packages/terminal/src/browser/base/terminal-service.ts
