+++ packages/terminal/src/browser/base/terminal-widget.ts
