+++ packages/terminal/src/browser/terminal-frontend-contribution.ts
@@ -34,8 +34,8 @@ import {
 import { TabBarToolbarContribution, TabBarToolbarRegistry } from '@theia/core/lib/browser/shell/tab-bar-toolbar';
 import { WidgetManager } from '@theia/core/lib/browser';
 import { TerminalKeybindingContexts } from './terminal-keybinding-contexts';
-import { TerminalService } from './terminal-service';
-import { TerminalWidgetOptions, TerminalWidget, TERMINAL_WIDGET_FACTORY_ID, TerminalWidgetFactoryOptions } from './terminal-widget';
+import { TerminalService } from './base/terminal-service';
+import { TerminalWidgetOptions, TerminalWidget, TERMINAL_WIDGET_FACTORY_ID, TerminalWidgetFactoryOptions } from './base/terminal-widget';
 import { UriAwareCommandHandler } from '@theia/core/lib/common/uri-command-handler';
 import { FileSystem } from '@theia/filesystem/lib/common';
 import URI from '@theia/core/lib/common/uri';
